Construction - fill, add method swap and fixed overflow

+++ nikiforov.alexander/S1/main.cpp
@@ -4,8 +4,8 @@ int main()
 {
   using namespace nikiforov;
 
-  List< std::pair< std::string, List< int > > > seqsPair;
-  Funcs< int > func;
+  List< std::pair< std::string, List< unsigned long long > > > seqsPair;
+  Funcs< unsigned long long > func;
   try
   {
     func.input_(std::cin, seqsPair);
